refactor(ctld): extract node/partition metadata into Node/ subdirectory
Move CranedStaticMeta, CranedRemoteMeta, CranedMeta, ResvMeta,
PartitionGlobalMeta, PartitionMeta from CtldPublicDefs.h into
Node/NodeDefs.h. Move CranedMetaContainer.{h,cpp} into Node/.
Also add build targets to CMakePresets.json via inherited _base preset.
