+++ src/CraneCtld/CraneCtld.cpp
@@ -29,7 +29,7 @@
 #include "Account/AccountManager.h"
 #include "Accounting/AccountMetaContainer.h"
 #include "Accounting/LicenseManager.h"
-#include "CranedMetaContainer.h"
+#include "Node/CranedMetaContainer.h"
 #include "CtldPublicDefs.h"
 #include "Database/DbClient.h"
 #include "Database/EmbeddedDbClient.h"
