+++ src/CraneCtld/CtldPublicDefs.cpp
@@ -23,23 +23,6 @@
 
 namespace Ctld {
 
-CranedRemoteMeta::CranedRemoteMeta(
-    const crane::grpc::CranedRemoteMeta& grpc_meta)
-    : dres_in_node(grpc_meta.dres_in_node()) {
-  this->sys_rel_info.name = grpc_meta.sys_rel_info().name();
-  this->sys_rel_info.release = grpc_meta.sys_rel_info().release();
-  this->sys_rel_info.version = grpc_meta.sys_rel_info().version();
-  this->craned_start_time =
-      absl::FromUnixSeconds(grpc_meta.craned_start_time().seconds());
-  this->system_boot_time =
-      absl::FromUnixSeconds(grpc_meta.system_boot_time().seconds());
-
-  this->network_interfaces.clear();
-  for (const auto& interface : grpc_meta.network_interfaces()) {
-    this->network_interfaces.emplace_back(interface);
-  }
-}
-
 PodMetaInJob::PodMetaInJob(const crane::grpc::PodJobAdditionalMeta& rhs)
     : name(rhs.name()),
       labels(rhs.labels().begin(), rhs.labels().end()),
